Handling the user flavors
