+++ README.md
@@ -12,13 +12,12 @@ npm run dev
 
 Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
 
+Reference from older application: https://github.com/aaronsisler/clients.lawndale-manor/blob/main/src/components/contact-form/contact-form.jsx
+
 ## Review or TODO Items
 
+- A way to trigger actions, such as logging in if the application wasn't ready yet
 - Figure out the Client | undefined issue...
-- react hook form
-- Example from older application: https://github.com/aaronsisler/clients.lawndale-manor/blob/main/src/components/contact-form/contact-form.jsx
-- zod is a form validation tool?
 - Can you generate "models" from a contract?
-- GitHub workflow
 - Should we use npm, yarn, pnpm, bun?
 - Getting a "Actions on Save" working such that the file and spacing/imports all format
