+++ src/app/common/header.tsx
@@ -1,31 +1,51 @@
 "use client";
 
+import { useEffect } from "react";
 import Link from "next/link";
-import { UserState, useUserStore } from "../user/user-store";
+
+import { useErrorStore, ErrorState } from "../common/error-store";
 import { User } from "../user/user";
+import { UserState, useUserStore } from "../user/user-store";
+import { UserRepository } from "../user/user-repository";
 
 const Header = () => {
   const user: User | undefined = useUserStore((state: UserState) => state.user);
 
+  const { setUser } = useUserStore((state: UserState) => state);
+  const { clearErrorMessage, setErrorMessage } = useErrorStore(
+    (state: ErrorState) => state
+  );
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        clearErrorMessage();
+        const user: User = await UserRepository.get();
+        setUser(user);
+      } catch (error: any) {
+        setErrorMessage(error);
+      }
+    };
+    fetchData();
+  }, [setUser, clearErrorMessage, setErrorMessage]);
+
   return (
-    <header className="pb-2.5">
-      <div className="">
-        <Link href="/">Home</Link>&nbsp;
-        <Link href="/client">Client</Link>&nbsp;
-        <Link href="/organizer">Organizer</Link>&nbsp;
-        <Link href="/location">Location</Link>
+    <header className="mb-2.5 flex bg-slate-400">
+      <div className="flex-grow-0 gap-10">
+        <Link className="mr-2" href="/">
+          Home
+        </Link>
+        <Link className="mr-2" href="/client">
+          Client
+        </Link>
+        <Link className="mr-2" href="/organizer">
+          Organizer
+        </Link>
+        <Link className="mr-2" href="/location">
+          Location
+        </Link>
       </div>
-      <div className="">
-        <select
-          onChange={(data: any) => {
-            console.log(data);
-          }}
-        >
-          {user?.clientIds?.map((clientId) => (
-            <option key={clientId}>{clientId}</option>
-          ))}
-        </select>
-      </div>
+      <div className="pl-2.5">{user?.name || "No user found"}</div>
     </header>
   );
 };
